requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
dj-database-url>=2.0
gunicorn>=21.0
whitenoise>=6.0
twilio>=8.0
anthropic>=0.30
resend>=0.6
python-dotenv>=1.0
apscheduler>=3.10
